+++ src/modules/core/hooks/useRedirectToFirstVault.ts
@@ -1,0 +1,36 @@
+import { Pages } from '@/modules';
+import { useUpdateSettingsRequest } from '@/modules/settings/hooks';
+import { useWorkspaceContext } from '@/modules/workspace/WorkspaceProvider';
+
+const useRedirectToFirstVault = () => {
+  const updateUserMutation = useUpdateSettingsRequest();
+  const {
+    workspaceInfos: {
+      handlers: { handleWorkspaceSelection },
+    },
+  } = useWorkspaceContext();
+
+  const handleRedirectToFirstVault = (
+    vaultId: string,
+    workspaceId: string,
+    userId: string,
+  ) => {
+    updateUserMutation.mutate({
+      id: userId,
+      first_login: false,
+    });
+
+    handleWorkspaceSelection(
+      workspaceId,
+      Pages.detailsVault({
+        vaultId: vaultId ?? '',
+        workspaceId: workspaceId ?? '',
+      }),
+    );
+  };
+
+  return {
+    handleRedirectToFirstVault,
+  };
+};
+export { useRedirectToFirstVault };
